refactor: ♻️ split bloated test files into directory structure (#177)
- agent_sessions_api.rs (615L) → agent_sessions/{crud,lifecycle,restart,outputs}.rs
- authorization_api.rs (586L) → authorization/{projects,tasks_members,owner_protection}.rs
- task_service.rs (992L) → task_service/{mod,tests}.rs
- TaskDetailModal.test.tsx (528L) → split into rendering + actions test files
- Extract shared auth helpers to common/mod.rs
- Extract shared frontend mock setup to __tests__/taskDetailModalSetup.ts
