+++ backend/tests/agent_sessions/main.rs
@@ -1,0 +1,7 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod crud;
+mod lifecycle;
+mod outputs;
+mod restart;
