+++ backend/tests/authorization/main.rs
@@ -1,0 +1,6 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod owner_protection;
+mod projects;
+mod tasks_members;
